+++ R/count-tally-SpatVector.R
@@ -2,10 +2,10 @@
 #'
 #' @description
 #' `count()` lets you quickly count the unique values of one or more variables:
-#' * `df |> count(a, b)` is roughly equivalent to
-#'   `df |> group_by(a, b) |> summarise(n = n())`.
-#' * `count()` is paired with `tally()`, a lower-level helper that is equivalent
-#'    to `df |> summarise(n = n())`.
+#' `df |> count(a, b)` is roughly equivalent to
+#' `df |> group_by(a, b) |> summarise(n = n())`. `count()` is paired with
+#' `tally()`, a lower-level helper that is equivalent to
+#' `df |> summarise(n = n())`.
 #'
 #' @export
 #' @rdname count.SpatVector