+++ R/pivot-wide-SpatVector.R
@@ -37,11 +37,11 @@
 #'
 #' @section Methods:
 #'
-#' Implementation of the **generic** [tidyr::pivot_wider()] function.
+#' Implementation of the **generic** [tidyr::pivot_wider()] method.
 #'
 #' ## `SpatVector`
 #'
-#' The geometry column has a sticky behavior. This means that the result would
+#' The geometry column has a sticky behaviour. This means that the result would
 #' have always the geometry of `data`.
 #'
 #' @examples