+++ R/slice-Spat.R
@@ -68,7 +68,7 @@
 #'
 #' @section Methods:
 #'
-#' Implementation of the **generic** [dplyr::slice()] function.
+#' Implementation of the **generic** [dplyr::slice()] method.
 #'
 #' ## `SpatRaster`
 #'
@@ -182,12 +182,12 @@ slice.SpatRaster <- function(
 }
 #' @export
 #' @rdname slice.Spat
-slice.SpatVector <- function(.data, ..., .preserve = FALSE) {
+slice.SpatVector <- function(.data, ..., .by = NULL, .preserve = FALSE) {
   # Use own method
   tbl <- as_tibble(.data)
   ind <- make_safe_index("tterra_index", tbl)
   tbl[[ind]] <- seq_len(nrow(tbl))
-  sliced <- dplyr::slice(tbl, ..., .preserve = .preserve)
+  sliced <- dplyr::slice(tbl, ..., .by = {{ .by }}, .preserve = .preserve)
 
   # Regenerate
   vend <- .data[as.integer(sliced[[ind]]), ]
@@ -227,13 +227,13 @@ slice_head.SpatRaster <- function(.data, ..., n, prop, .keep_extent = FALSE) {
 
 #' @export
 #' @rdname slice.Spat
-slice_head.SpatVector <- function(.data, ..., n, prop) {
+slice_head.SpatVector <- function(.data, ..., n, prop, by = NULL) {
   # Use own method
   tbl <- as_tibble(.data)
   ind <- make_safe_index("tterra_index", tbl)
   tbl[[ind]] <- seq_len(nrow(tbl))
 
-  sliced <- dplyr::slice_head(tbl, ..., n = n, prop = prop)
+  sliced <- dplyr::slice_head(tbl, ..., n = n, prop = prop, by = {{ by }})
 
   # Regenerate
   vend <- .data[as.integer(sliced[[ind]]), ]
@@ -274,13 +274,13 @@ slice_tail.SpatRaster <- function(.data, ..., n, prop, .keep_extent = FALSE) {
 
 #' @export
 #' @rdname slice.Spat
-slice_tail.SpatVector <- function(.data, ..., n, prop) {
+slice_tail.SpatVector <- function(.data, ..., n, prop, by = NULL) {
   # Use own method
   tbl <- as_tibble(.data)
   ind <- make_safe_index("tterra_index", tbl)
   tbl[[ind]] <- seq_len(nrow(tbl))
 
-  sliced <- dplyr::slice_tail(tbl, ..., n = n, prop = prop)
+  sliced <- dplyr::slice_tail(tbl, ..., n = n, prop = prop, by = {{ by }})
 
   # Regenerate
   vend <- .data[as.integer(sliced[[ind]]), ]
@@ -358,6 +358,7 @@ slice_min.SpatVector <- function(
   ...,
   n,
   prop,
+  by = NULL,
   with_ties = TRUE,
   na_rm = FALSE
 ) {
@@ -368,11 +369,11 @@ slice_min.SpatVector <- function(
 
   sliced <- dplyr::slice_min(
     tbl,
-    ...,
     order_by = {{ order_by }},
     ...,
     n = n,
     prop = prop,
+    by = {{ by }},
     with_ties = with_ties,
     na_rm = na_rm
   )
@@ -453,6 +454,7 @@ slice_max.SpatVector <- function(
   ...,
   n,
   prop,
+  by = NULL,
   with_ties = TRUE,
   na_rm = FALSE
 ) {
@@ -468,6 +470,7 @@ slice_max.SpatVector <- function(
     ...,
     n = n,
     prop = prop,
+    by = {{ by }},
     with_ties = with_ties,
     na_rm = na_rm
   )
@@ -541,6 +544,7 @@ slice_sample.SpatVector <- function(
   ...,
   n,
   prop,
+  by = NULL,
   weight_by = NULL,
   replace = FALSE
 ) {
@@ -549,7 +553,15 @@ slice_sample.SpatVector <- function(
   ind <- make_safe_index("tterra_index", tbl)
   tbl[[ind]] <- seq_len(nrow(tbl))
 
-  sliced <- dplyr::slice_sample(tbl, ..., n = n, prop = prop, replace = replace)
+  sliced <- dplyr::slice_sample(
+    tbl,
+    ...,
+    n = n,
+    prop = prop,
+    by = {{ by }},
+    weight_by = {{ weight_by }},
+    replace = replace
+  )
 
   # Regenerate
   vend <- .data[as.integer(sliced[[ind]]), ]